dice-protocols: tcat: extension: code refactoring to obsolete RouterEntryData
In former commits, RouterEntryData is not used by runtime.

This commit enclose the usage of RouterEntryData in protocol crate.
