+++ libs/dice/protocols/src/tcat/extension.rs
@@ -184,79 +184,11 @@ impl Default for DstBlkId {
     }
 }
 
-impl From<u8> for DstBlkId {
-    fn from(val: u8) -> Self {
-        match val {
-            0 => Self::Aes,
-            1 => Self::Adat,
-            2 => Self::MixerTx0,
-            3 => Self::MixerTx1,
-            4 => Self::Ins0,
-            5 => Self::Ins1,
-            10 => Self::ArmApbAudio,
-            11 => Self::Avs0,
-            12 => Self::Avs1,
-            _ => Self::Reserved(val),
-        }
-    }
-}
-
-impl From<DstBlkId> for u8 {
-    fn from(id: DstBlkId) -> Self {
-        match id {
-            DstBlkId::Aes => 0,
-            DstBlkId::Adat => 1,
-            DstBlkId::MixerTx0 => 2,
-            DstBlkId::MixerTx1 => 3,
-            DstBlkId::Ins0 => 4,
-            DstBlkId::Ins1 => 5,
-            DstBlkId::ArmApbAudio => 10,
-            DstBlkId::Avs0 => 11,
-            DstBlkId::Avs1 => 12,
-            DstBlkId::Reserved(val) => val,
-        }
-    }
-}
-
 /// The structure to represent destination block.
 #[derive(Default, Debug, Clone, Copy, Eq, PartialEq)]
 pub struct DstBlk {
     pub id: DstBlkId,
     pub ch: u8,
-}
-
-impl DstBlk {
-    pub const ID_MASK: u8 = 0xf0;
-    pub const ID_SHIFT: usize = 4;
-    pub const CH_MASK: u8 = 0x0f;
-    pub const CH_SHIFT: usize = 0;
-}
-
-impl From<u8> for DstBlk {
-    fn from(val: u8) -> Self {
-        DstBlk {
-            id: DstBlkId::from((val & DstBlk::ID_MASK) >> DstBlk::ID_SHIFT),
-            ch: (val & DstBlk::CH_MASK) >> DstBlk::CH_SHIFT,
-        }
-    }
-}
-
-impl From<DstBlk> for u8 {
-    fn from(blk: DstBlk) -> Self {
-        (u8::from(blk.id) << DstBlk::ID_SHIFT) | blk.ch
-    }
-}
-
-impl Ord for DstBlk {
-    fn cmp(&self, other: &Self) -> Ordering {
-        u8::from(*self).cmp(&u8::from(*other))
-    }
-}
-
-impl PartialOrd for DstBlk {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(u8::from(*self).cmp(&u8::from(*other)))
-    }
 }
 
 /// The enumeration to represent ID of source block.
@@ -280,40 +212,6 @@ impl Default for SrcBlkId {
     }
 }
 
-impl From<u8> for SrcBlkId {
-    fn from(val: u8) -> Self {
-        match val {
-            0 => Self::Aes,
-            1 => Self::Adat,
-            2 => Self::Mixer,
-            4 => Self::Ins0,
-            5 => Self::Ins1,
-            10 => Self::ArmAprAudio,
-            11 => Self::Avs0,
-            12 => Self::Avs1,
-            15 => Self::Mute,
-            _ => Self::Reserved(val),
-        }
-    }
-}
-
-impl From<SrcBlkId> for u8 {
-    fn from(id: SrcBlkId) -> Self {
-        match id {
-            SrcBlkId::Aes => 0,
-            SrcBlkId::Adat => 1,
-            SrcBlkId::Mixer => 2,
-            SrcBlkId::Ins0 => 4,
-            SrcBlkId::Ins1 => 5,
-            SrcBlkId::ArmAprAudio => 10,
-            SrcBlkId::Avs0 => 11,
-            SrcBlkId::Avs1 => 12,
-            SrcBlkId::Mute => 15,
-            SrcBlkId::Reserved(val) => val,
-        }
-    }
-}
-
 /// The structure to represent source block.
 #[derive(Default, Debug, Clone, Copy, Eq, PartialEq)]
 pub struct SrcBlk {
@@ -321,78 +219,12 @@ pub struct SrcBlk {
     pub ch: u8,
 }
 
-impl SrcBlk {
-    pub const ID_MASK: u8 = 0xf0;
-    pub const ID_SHIFT: usize = 4;
-    pub const CH_MASK: u8 = 0x0f;
-    pub const CH_SHIFT: usize = 0;
-}
-
-impl From<u8> for SrcBlk {
-    fn from(val: u8) -> Self {
-        SrcBlk {
-            id: SrcBlkId::from((val & SrcBlk::ID_MASK) >> SrcBlk::ID_SHIFT),
-            ch: (val & SrcBlk::CH_MASK) >> SrcBlk::CH_SHIFT,
-        }
-    }
-}
-
-impl From<SrcBlk> for u8 {
-    fn from(blk: SrcBlk) -> Self {
-        (u8::from(blk.id) << SrcBlk::ID_SHIFT) | blk.ch
-    }
-}
-
-impl Ord for SrcBlk {
-    fn cmp(&self, other: &Self) -> Ordering {
-        u8::from(*self).cmp(&u8::from(*other))
-    }
-}
-
-impl PartialOrd for SrcBlk {
-    fn partial_cmp(&self, other: &Self) -> Option<Ordering> {
-        Some(u8::from(*self).cmp(&u8::from(*other)))
-    }
-}
-
-/// The alternative type of data for router entry.
-pub type RouterEntryData = [u8;RouterEntry::SIZE];
-
 /// The structure to represent entry of route.
 #[derive(Default, Debug, Clone, Copy, Eq, PartialEq)]
 pub struct RouterEntry {
     pub dst: DstBlk,
     pub src: SrcBlk,
     pub peak: u16,
-}
-
-impl RouterEntry {
-    const SIZE: usize = 4;
-
-    pub const SRC_OFFSET: usize = 2;
-    pub const DST_OFFSET: usize = 3;
-}
-
-impl From<&RouterEntryData> for RouterEntry {
-    fn from(raw: &RouterEntryData) -> Self {
-        let mut doublet = [0;2];
-        doublet.copy_from_slice(&raw[..2]);
-        RouterEntry {
-            dst: raw[Self::DST_OFFSET].into(),
-            src: raw[Self::SRC_OFFSET].into(),
-            peak: u16::from_be_bytes(doublet),
-        }
-    }
-}
-
-impl From<&RouterEntry> for RouterEntryData {
-    fn from(entry: &RouterEntry) -> RouterEntryData {
-        let mut raw = RouterEntryData::default();
-        raw.copy_from_slice(&entry.peak.to_be_bytes());
-        raw[RouterEntry::SRC_OFFSET] = entry.src.into();
-        raw[RouterEntry::DST_OFFSET] = entry.dst.into();
-        raw
-    }
 }
 
 /// The structure to represent entry of stream format.
@@ -473,7 +305,6 @@ impl From<FormatEntry> for FormatEntryData {
 mod test {
     use super::Section;
     use super::ExtensionSections;
-    use super::{DstBlk, SrcBlk, DstBlkId, SrcBlkId};
     use super::{FormatEntry, FormatEntryData, AC3_CHANNELS};
 
     use std::convert::TryFrom;
@@ -500,24 +331,6 @@ mod test {
             application: Section{offset: 0x04, size: 0x00},
         };
         assert_eq!(space, ExtensionSections::from(&raw[..]));
-    }
-
-    #[test]
-    fn dst_blk_from() {
-        let blk = DstBlk {
-            id: DstBlkId::ArmApbAudio,
-            ch: 0x04,
-        };
-        assert_eq!(blk, DstBlk::from(u8::from(blk)));
-    }
-
-    #[test]
-    fn src_blk_from() {
-        let blk = SrcBlk {
-            id: SrcBlkId::ArmAprAudio,
-            ch: 0x04,
-        };
-        assert_eq!(blk, SrcBlk::from(u8::from(blk)));
     }
 
     #[test]
